Rollback of commit 67b4d5250edcefa7220e928e529b1f385e2dc464.
*** Reason for rollback ***

Memory regression: []

*** Original change description ***

Pass direct jars to Jack.

The R class generator produces a compiled jar, not a java source file.
This means all android libraries/binaries with resources will need to
convert this jar to jack to ensure that they have access to jars.

This change will not fix Jack builds, but it is a necessary step toward
fixing them.

--
PiperOrigin-RevId: 141091215
MOS_MIGRATED_REVID=141091215